chore: Cleaned up code for groups tree
