+++ app/assets/stylesheets/application.tailwind.css
@@ -279,89 +279,127 @@ body {
   }
 
   /*GROUPS LIST TREE*/
-  .groups-list-tree-container
-    > .group-list-tree
-    > .group-row.has-children:first-child {
-    @apply border-t-0;
+  .groups-list-tree-container > .group-list-tree > .group-row.has-children:first-child {
+    border-top: 0;
   }
 
   .group-list-tree .folder-toggle-wrap {
     font-size: 0;
-    @apply flex-shrink-0;
+    flex-shrink: 0;
   }
-
   .group-list-tree .folder-toggle-wrap span {
-    @apply text-sm;
+    font-size: 12px;
   }
 
-  .group-list-tree .group-row:not(.has-children) .folder-caret {
-    @apply opacity-0;
+  .group-list-tree .item-type-icon {
+    margin-top: 2px;
+    width: 20px;
+  }
+
+  .group-list-tree > .group-row:not(.has-children) .folder-caret {
+    opacity: 0;
   }
 
   .group-list-tree .group-list-tree {
-    @apply mb-0 ml-8 relative;
+    margin-bottom: 0;
+    margin-left: 30px;
+    position: relative;
   }
 
   .group-list-tree .group-list-tree::before {
-    content: "";
+    content: '';
+    display: block;
+    width: 0;
+    position: absolute;
     top: 5px;
     bottom: 0;
     left: -16px;
-    @apply absolute border-l-2 border-slate-200 dark:border-slate-600 block w-0;
+    @apply border-l-2 border-slate-200 dark:border-slate-600;
   }
 
   .group-list-tree .group-list-tree .group-row {
-    @apply relative;
+    position: relative;
   }
-
   .group-list-tree .group-list-tree .group-row::before {
-    content: "";
+    content: '';
+    display: block;
     width: 10px;
     height: 0;
+    position: absolute;
     top: 30px;
     left: -16px;
-    @apply block border-t-2 border-slate-200 dark:border-slate-600 absolute;
+    @apply border-t-2 border-slate-200 dark:border-slate-600;
   }
-
   .group-list-tree .group-list-tree .group-row:last-child::before {
+    background: white;
     height: auto;
     top: 30px;
     bottom: 0;
-    @apply bg-white;
   }
 
+  .group-list-tree .group-list-tree .group-row.being-removed {
+    opacity: 0.5;
+  }
   .group-list-tree .group-row {
-    @apply p-0;
+    padding: 0;
   }
-
   .group-list-tree .group-row.has-children {
-    @apply border-t-0;
+    border-top: 0;
   }
-
   .group-list-tree .group-row:first-child {
-    @apply border-t-2 border-slate-50 dark:border-slate-700;
+    @apply border-t border-slate-200 dark:border-slate-600;
   }
 
-  .group-list-tree .group-row-contents.has_children:hover {
-    @apply bg-slate-50 dark:bg-slate-900 cursor-pointer;
+  .group-list-tree .group-row.has-children > .group-row-contents:hover {
+    @apply cursor-pointer border-slate-50 dark:border-slate-600 bg-slate-50 dark:bg-slate-600;
   }
 
-  .group-list-tree .group-row-contents .group-text-container,
-  .group-list-tree .group-row-contents .group-text {
-    min-width: 0; /* allows for truncated text within flex children */
+  .group-list-tree .group-row-contents .group-text-container, .group-list-tree .group-row-contents .group-text {
+    min-width: 0;
   }
-
   .group-list-tree .group-row-contents .group-text {
     flex-basis: 100%;
   }
 
-  .group-list-tree .group-row-contents .title {
-    @apply text-sm font-medium truncate text-slate-800;
+  .group-list-tree .group-row-contents .avatar-container {
+    flex-shrink: 0;
   }
 
-  .group-list-tree .group-row-contents .description {
-    @apply text-sm truncate text-slate-600;
+  .group-list-tree .group-row-contents .avatar-container > a {
+    width: 100%;
+    text-decoration: none;
   }
+  .group-list-tree .group-row-contents .title {
+    font-size: 16px;
+  }
+
+  .group-list-tree .group-row-contents.has-more-items {
+    display: block;
+    padding: 20px 10px;
+  }
+
+  .group-list-tree .group-row-contents .description p {
+    max-width: 100%;
+  }
+
+  .group-list-tree .group-row-contents .stats {
+    position: relative;
+    line-height: normal;
+    text-align: right;
+    flex-shrink: 0;
+  }
+
+  .group-list-tree .group-row-contents .stats > span {
+    display: inline-flex;
+    align-items: center;
+    height: 16px;
+    min-width: 30px;
+  }
+
+  .group-list-tree .group-row-contents .stats .stat-value {
+    margin: 2px 0 0 5px;
+  }
+
 }
 
 .pagy-nav.pagination {
