--- app/components/group_list/group_row_component.html.erb
@@ -1,65 +1,0 @@
-<% if group.children.count > 0 %>
-  <li
-  id="<%= dom_id(group) %>"
-  class="group-row"
-  data-controller="groups--row"
->
-    <div class="group-row-contents has_children flex items-center py-2 pr-3"
-         data-action="click->groups--row#toggle" role="button" aria-label="<%= t(:'dashboard.groups.index.row_aria_label') %>">
-      <div class="folder-toggle-wrap flex items-center w-3">
-        <%= link_to public_send(path, path_args.merge({ parent_id: group.id, collapse: !collapse })), data: { turbo_stream: '', "groups--row-target": "toggle" },
-                    aria: { label: t(:'dashboard.groups.index.row_aria_label') } do %>
-          <%= viral_icon(
-                name: collapse ? :chevron_right : :chevron_down,
-                classes: "h-3 w-3 inline-block"
-              ) %>
-        <% end %>
-      </div>
-<% else %>
-  <li>
-    <div class="group-row-contents flex items-center py-2 pr-3">
-      <div class="folder-toggle-wrap flex items-center w-3">
-
-      </div>
-<% end %>
-    <div class="w-10 p-2 mr-3">
-      <%= viral_avatar(
-        name: group.name,
-        size: :small,
-        colour_string: "#{group.name}-#{group.id}",
-        data: {
-          turbo: false
-        },
-        url: group_path(group)
-      ) %>
-    </div>
-    <div class="group-text-container flex items-center flex-grow">
-      <div class="group-text grow">
-        <div class="flex items-center flex-wrap mr-3 title">
-          <%= link_to group.name, group_path(id: group.to_param) %>
-        </div>
-        <div class="description">
-          <p><%= group.description %></p>
-        </div>
-      </div>
-      <div class="metadata grow flex-shrink-0 flex-wrap justify-between">
-        <span class="flex items-center"><%= viral_icon(name: "squares_2x2", color: :subdued, classes: "h-5 w-5") %><%= group.children.count %></span>
-      </div>
-    </div>
-  </div>
-<% unless collapse %>
-    <%= render GroupList::GroupListTreeComponent.new(groups: group.children, path:) %>
-    <% if group.children.length > 10 %>
-      <div class="group-row-contents flex items-center py-2 pr-3">
-        <div class="group-text-container flex items-center flex-grow">
-          <div class="group-text grow">
-            <div class="flex items-center flex-wrap mr-3 title">
-              <%= link_to "View all #{group.children.count} subgroups",
-                          public_send(path, path_args.merge({ parent_id: group.id })) %>
-            </div>
-          </div>
-        </div>
-      </div>
-    <% end %>
-  <% end %>
-</li>
