+++ app/components/groups_list/group_row/row_contents_component.html.erb
@@ -1,0 +1,38 @@
+<%= link_to public_send(@path, @path_args.merge({ parent_id: @group.id, collapse: !@collapsed })), data: { turbo_stream: '', "groups--row-target": "toggle" },
+            aria: { label: t(:'dashboard.groups.index.row_aria_label') }, class: "folder-toggle-wrap mr-2 flex items-center" do %>
+  <span class="folder-caret inline-block text-slate-400">
+      <%= viral_icon(name: :chevron_right, classes: 'h-3 w-3') %>
+    </span>
+  <span class="item-type-icon text-slate-400">
+      <%= viral_icon(name: :squares_2x2, classes: 'h-4 w-4') %>
+    </span>
+<% end %>
+
+<%= viral_avatar(
+      name: @group.name,
+      size: :small,
+      colour_string: "#{@group.name}-#{@group.id}",
+      data: {
+        turbo: false
+      },
+      url: group_path(@group)
+    ) %>
+
+<div class="group-text-container flex grow items-center ml-2">
+  <div class="group-text flex flex-col">
+    <div class="group-text grow shrink">
+      <div class="flex items-center flex-wrap title font-semibold mr-3">
+        <%= link_to @group.name, group_path(@group) %>
+      </div>
+      <div class="description">
+        <p><%= @group.description %></p>
+      </div>
+    </div>
+  </div>
+</div>
+
+<div class="metadata flex flex-none shrink-0 flex-wrap justify-between">
+  <div class="stats text-slate-500 group-stats mt-2 flex items-center">
+    <span class="flex items-center" title="<%= t(:'dashboard.groups.index.subgroups') %>"><%= viral_icon(name: "squares_2x2", color: :subdued, classes: "h-5 w-5") %><%= @group.children.count %></span>
+  </div>
+</div>
