+++ app/components/groups_list/group_row/with_children_component.html.erb
@@ -1,0 +1,9 @@
+<li id="<%= dom_id(@group) %>" class="group-row has-children <%= @group.description.present? ? 'has-description' : nil %>" data-controller="groups--row" %>
+  <div class="group-row-contents py-3 pr-3 flex items-center"
+       data-action="click->groups--row#toggle" role="button" aria-label="<%= t(:'dashboard.groups.index.row_aria_label') %>">
+    <%= render GroupsList::GroupRow::RowContentsComponent.new(group: @group, path: @path, path_args: @path_args, collapsed: @collapsed) %>
+  </div>
+  <% unless @collapsed %>
+    <%=  render GroupsList::GroupListTreeComponent.new(groups: @group.children, path: @path, path_args: @path_args)  %>
+  <%  end %>
+</li>
