+++ app/components/groups_list/group_row/with_children_component.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module GroupsList
+  module GroupRow
+    class WithChildrenComponent < ViewComponent::Base
+      def initialize(group:, path: nil, path_args: {}, collapsed: false)
+        @group = group
+        @path = path
+        @path_args = path_args
+        @collapsed = collapsed
+      end
+    end
+  end
+end
