+++ app/components/groups_list/group_row/without_children_component.html.erb
@@ -1,0 +1,5 @@
+<li class="group-row <%= @group.description.present? ? "has-description" : nil %>">
+  <div class="group-row-contents py-3 pr-3 flex items-center">
+    <%= render GroupsList::GroupRow::RowContentsComponent.new(group: @group, path: @path, path_args: @path_args) %>
+  </div>
+</li>
