+++ app/components/groups_list/group_row/without_children_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class GroupsList::GroupRow::WithoutChildrenComponent < ViewComponent::Base
+  def initialize(group:, path: nil, path_args: {})
+    @group = group
+    @path = path
+    @path_args = path_args
+  end
+end
