+++ app/views/dashboard/groups/index.html.erb
@@ -10,9 +10,5 @@
   <% end %>
 <% end %>
 <div class="bg-white dark:bg-gray-800">
-  <div class="flex flex-col">
-    <div class="overflow-x-auto">
-      <%= render GroupListComponent.new(groups: @groups, path: "dashboard_groups_path") %>
-    </div>
-  </div>
+  <%= render GroupsListTreeContainerComponent.new(groups: @groups, path: "dashboard_groups_path") %>
 </div>
