+++ config/locales/en.yml
@@ -290,6 +290,7 @@ en:
       index:
         more_items: "%{count} more items"
         row_aria_label: "See %{group_name} subgroups"
+        subgroups: Subgroups
   profiles:
     sidebar:
       header: User Settings
